npm vulns and reworked webpack - still fix build
