+++ package.json
@@ -8,36 +8,36 @@
     "test:watch": "npm run test -- --watch",
     "build": "NODE_ENV=production webpack",
     "start": "node ./src/app.js",
-    "dev": "webpack-dev-server"
+    "dev": "webpack serve --open"
   },
   "keywords": [],
   "author": "Kyle Burke <k23burke@gmail.com>",
   "license": "ISC",
   "devDependencies": {
-    "babel-core": "^6.7.2",
-    "babel-loader": "^6.2.4",
-    "babel-preset-es2015": "^6.6.0",
-    "babel-preset-react": "^6.5.0",
+    "@babel/core": "^7.13.1",
+    "@babel/preset-env": "^7.13.5",
+    "@babel/preset-react": "^7.12.13",
+    "babel-loader": "^8.2.2",
     "chai": "^3.5.0",
     "chai-immutable": "^1.5.3",
     "clean-webpack-plugin": "^0.1.11",
-    "css-loader": "^0.23.1",
-    "file-loader": "^0.8.5",
-    "image-webpack-loader": "^1.7.0",
+    "css-loader": "^5.0.2",
+    "file-loader": "^6.2.0",
+    "image-webpack-loader": "^7.0.1",
     "jsdom": "^8.1.0",
-    "mocha": "^2.4.5",
-    "node-sass": "^3.4.2",
+    "mocha": "^8.3.0",
+    "node-sass": "^5.0.0",
     "react-hot-loader": "^1.3.0",
-    "sass-loader": "^3.2.0",
+    "sass-loader": "^11.0.1",
     "style-loader": "^0.13.0",
-    "url-loader": "^0.5.7",
-    "webpack": "^1.12.15",
-    "webpack-dev-server": "^1.14.1"
+    "url-loader": "^4.1.1",
+    "webpack": "^5.24.0",
+    "webpack-cli": "^4.5.0",
+    "webpack-dev-server": "^3.11.2"
   },
   "babel": {
     "presets": [
-      "es2015",
-      "react"
+      "@babel/preset-react"
     ]
   },
   "dependencies": {
@@ -48,16 +48,17 @@
     "lodash": "^4.16.2",
     "moment": "^2.15.1",
     "node-uuid": "^1.4.7",
-    "react": "^15.3.2",
+    "react": "^17.0.1",
     "react-addons-pure-render-mixin": "^15.3.2",
-    "react-dom": "^15.3.2",
+    "react-dom": "^17.0.1",
     "react-redux": "^4.4.5",
     "react-router": "^2.8.1",
     "react-router-scroll": "^0.3.2",
     "redux": "^3.6.0",
-    "redux-api-middleware": "^1.0.2",
+    "redux-api-middleware": "^3.2.1",
     "redux-logger": "^2.6.1",
     "redux-thunk": "^2.1.0",
+    "sass": "^1.32.8",
     "whatwg-fetch": "^0.11.0"
   }
 }
