--- src/components/Bio.jsx
@@ -1,9 +1,0 @@
-export default () => {
-  return (
-    <div className='bio-container'>
-      <p>A trained software engineer, practiced disc jockey, and free form dabbler of synthesizers.<br/>Interested in the intersection between sonic experiences and technology. Currently in a super-group building a community engagement protocol at <a href="https://mochi.game/" target="_blank" rel="noopener noreferrer">Mochi.game</a>. Interested in joining? <a href="https://mochi.club" target="_blank" rel="noopener noreferrer">Apply for membership here.</a></p>
-      <p></p>
-    </div>
-
-  )
-}
